+++ src/config/config_manager.cc
@@ -41,6 +41,7 @@
 #include <langinfo.h>
 #endif
 
+#include "config/result/autoscan.h"
 #include "config/result/client_config.h"
 #include "config/result/transcoding.h"
 #include "config/setup/config_setup_path.h"
@@ -48,7 +49,6 @@
 #include "config_option_enum.h"
 #include "config_options.h"
 #include "config_setup.h"
-#include "content/autoscan.h"
 #include "database/database.h"
 #include "util/string_converter.h"
 #include "util/tools.h"
