+++ src/config/setup/config_setup_autoscan.cc
@@ -29,10 +29,10 @@
 #include "config/config_definition.h"
 #include "config/config_option_enum.h"
 #include "config/config_options.h"
+#include "config/result/autoscan.h"
 #include "config/setup/config_setup_enum.h"
 #include "config/setup/config_setup_path.h"
 #include "config/setup/config_setup_time.h"
-#include "content/autoscan.h"
 #include "content/autoscan_list.h"
 
 #ifdef HAVE_INOTIFY
