+++ src/content/import_service.cc
@@ -29,9 +29,9 @@
 #include <fmt/chrono.h>
 #include <regex>
 
-#include "autoscan.h"
 #include "cds/cds_container.h"
 #include "cds/cds_item.h"
+#include "config/result/autoscan.h"
 #include "content_manager.h"
 #include "database/database.h"
 #include "layout/builtin_layout.h"
