+++ src/content/layout/builtin_layout.cc
@@ -39,8 +39,8 @@
 #include "cds/cds_container.h"
 #include "cds/cds_item.h"
 #include "config/config.h"
+#include "config/result/autoscan.h"
 #include "config/result/box_layout.h"
-#include "content/autoscan.h"
 #include "content/content_manager.h"
 #include "metadata/metadata_handler.h"
 #include "util/string_converter.h"
