+++ src/content/onlineservice/curl_online_service.cc
@@ -29,7 +29,7 @@
 
 #include "cds/cds_item.h"
 #include "config/config.h"
-#include "content/autoscan.h"
+#include "config/result/autoscan.h"
 #include "content/content_manager.h"
 #include "content/layout/layout.h"
 #include "database/database.h"
