+++ src/content/scripting/script.cc
@@ -41,11 +41,11 @@
 #include "cds/cds_container.h"
 #include "cds/cds_item.h"
 #include "config/config_definition.h"
+#include "config/result/autoscan.h"
 #include "config/result/box_layout.h"
 #include "config/setup/config_setup_autoscan.h"
 #include "config/setup/config_setup_boxlayout.h"
 #include "config/setup/config_setup_dictionary.h"
-#include "content/autoscan.h"
 #include "content/content_manager.h"
 #include "database/database.h"
 #include "js_functions.h"
