+++ src/database/sql_database.cc
@@ -42,8 +42,8 @@
 #include "cds/cds_item.h"
 #include "config/config_manager.h"
 #include "config/config_setup.h"
+#include "config/result/autoscan.h"
 #include "config/result/dynamic_content.h"
-#include "content/autoscan.h"
 #include "content/autoscan_list.h"
 #include "metadata/metadata_handler.h"
 #include "search_handler.h"
