+++ src/server.cc
@@ -49,7 +49,7 @@
 #include "action_request.h"
 #include "config/config_manager.h"
 #include "config/config_option_enum.h"
-#include "content/autoscan.h"
+#include "config/result/autoscan.h"
 #include "content/content_manager.h"
 #include "database/database.h"
 #include "device_description_handler.h"
