+++ src/web/config_load.cc
@@ -32,6 +32,7 @@
 #include "config/config_definition.h"
 #include "config/config_option_enum.h"
 #include "config/config_setup.h"
+#include "config/result/autoscan.h"
 #include "config/result/box_layout.h"
 #include "config/result/client_config.h"
 #include "config/result/directory_tweak.h"
@@ -42,7 +43,6 @@
 #include "config/setup/config_setup_dictionary.h"
 #include "config/setup/config_setup_enum.h"
 #include "config/setup/config_setup_vector.h"
-#include "content/autoscan.h"
 #include "content/content_manager.h"
 #include "database/database.h"
 #include "metadata/metadata_handler.h"
