+++ src/web/directories.cc
@@ -36,7 +36,7 @@
 
 #include <array>
 
-#include "content/autoscan.h"
+#include "config/result/autoscan.h"
 #include "content/content_manager.h"
 #include "database/database.h"
 #include "util/string_converter.h"
